+++ CHANGES.markdown
@@ -3,6 +3,7 @@
   - Do not print empty lines before/after spec groups
   - Do not pass position to `Formatte.exampleGroupStarted` (we can not support
     this with the upcomming cleanup actions #188)
+  - Deprecate `Test.Hspec.Formatters.newParagraph`
 
 ## Changes in 1.11.4
   - Make test suite independent from QuickCheck seed (see #187)